+++ MatchRetriever/Exceptions/MatchNotFoundException.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Runtime.Serialization;
+
+namespace MatchRetriever.Exceptions
+{
+    public class MatchNotFoundException : Exception
+    {
+        public MatchNotFoundException()
+        {
+        }
+
+        public MatchNotFoundException(string message) : base(message)
+        {
+        }
+    }
+}