+++ MatchRetriever/Models/PlayerInMatchModel.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace MatchRetriever.Models
+{
+    public class PlayerInMatchModel
+    {
+        public List<long> Players { get; set; }
+    }
+}